backend/ai/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
google-generativeai